+++ src/syntax.jl
@@ -155,162 +155,3 @@ function analyze_syntax(arg)
     end
     return table
 end
-
-#####
-##### Count lines
-#####
-
-
-# We can't do this at the level of SyntaxNode's because we've lost whitespace & comments already
-
-# function count_lines(tree)
-#     items = PostOrderDFS(tree)
-#     d = Dict{Int, Vector{Any}}()
-#     for it in items
-#         node = it.node
-#         line, col = JuliaSyntax.source_location(node.source, node.position)
-#         v = get!(Vector{Any}, d, line)
-#         push!(v, kind(raw))
-#     end
-
-#     for k in sort!(collect(keys(d)))
-#         println(k, "\t|\t", d[k])
-#     end
-#     return d
-# end
-
-# So we will go to GreenNode's
-
-# Avoid piracy by defining AbstractTrees definitions on a wrapper
-struct GreenNodeWrapper2
-    node::JuliaSyntax.GreenNode
-    source::JuliaSyntax.SourceFile
-end
-
-function AbstractTrees.children(wrapper::GreenNodeWrapper2)
-    return map(n -> GreenNodeWrapper2(n, wrapper.source), JuliaSyntax.children(wrapper.node))
-end
-
-function parse_green_one(file_path)
-    file = read(file_path, String)
-    @debug(string("Parsing ", file_path))
-    parsed = JuliaSyntax.parse(JuliaSyntax.GreenNode, file; ignore_trivia=false)
-    return GreenNodeWrapper2(parsed[1], JuliaSyntax.SourceFile(file; filename=basename(file_path)))
-end
-
-# Module to make it easier w/r/t/ import clashes
-module CategorizeLines
-
-using JuliaSyntax: GreenNode, is_trivia, haschildren, is_error, children, span, SourceFile, source_location
-# Pretty printing
-function categorize_lines!(d, node, source, nesting=0, pos=1)
-    starting_line_number, col = source_location(source, pos)
-    v = get!(Vector{Any}, d, starting_line_number)
-
-    is_leaf = !haschildren(node)
-    if !is_leaf
-        new_nesting = nesting + 1
-        p = pos
-        for x in children(node)
-            categorize_lines!(d, x, source, new_nesting, p)
-            p += x.span
-        end
-        ending_line_number, _ = source_location(source, p)
-    else
-        ending_line_number = starting_line_number
-    end
-    push!(v, (; starting_line_number, ending_line_number, is_error=is_error(node), is_leaf, is_trivia=is_trivia(node), nesting, summary=summary(node)))
-    return nothing
-end
-end
-
-using .CategorizeLines
-
-
-# TODO:
-# Handle `@doc` calls?
-# What about inline comments #= comment =#?
-# Can a docstring not start at the beginning of a line?
-# Can there be multiple string nodes on the same line as a docstring?
-
-function identify_lines!(d2, v)
-    if v[1].summary == "Comment"
-        d2[v[1].starting_line_number] = "Comment"
-    elseif v[1].summary == "NewlineWs"
-        d2[v[1].starting_line_number] = "Blank"
-    elseif v[1].summary == "core_@doc"
-            idx = findfirst(x -> x.summary == "string", v)
-            for line in v[1].starting_line_number:v[idx].ending_line_number
-                d2[line] = "Docstring"
-            end
-    else
-        # Don't overwrite e.g. docstrings
-        if !haskey(d2, v[1].starting_line_number)
-            str = join((x.summary for x in v), ", ")
-            d2[v[1].starting_line_number] = "Code ($str)"
-        end
-    end
-end
-
-categorize_lines(path::AbstractString; kw...) = categorize_lines(parse_green_one(path); kw...)
-function categorize_lines(node::GreenNodeWrapper2; show=false)
-    d = Dict{Int, Vector{Any}}()
-    CategorizeLines.categorize_lines!(d, node.node, node.source)
-    d2 = Dict{Int, String}()
-
-    for k in sort!(collect(keys(d)))
-        identify_lines!(d2, d[k])
-    end
-
-    if show
-        for k in sort!(collect(keys(d2)))
-            println(k, "\t|\t", d2[k])
-        end
-    end
-    return d2
-end
-
-function _count_lines!(counts, node::GreenNodeWrapper2)
-    cats = categorize_lines(node; show=false)
-    for v in values(cats)
-        if startswith(v, "Code")
-            counts["Code"] += 1
-        else
-            counts[v] += 1
-        end
-    end
-    return nothing
-end
-
-function count_julia_loc(dir)
-    table = LoCTableEltype[]
-    for path in readdir(dir; join=true)
-        counts = Dict{String, Int}("Comment" => 0,
-                                   "Blank" => 0,
-                                   "Code" => 0,
-                                   "Docstring" => 0)
-        if isfile(path)
-            endswith(path, ".jl") || continue
-            node = parse_green_one(path)
-            _count_lines!(counts, node)
-            n_files = 1
-        else
-            n_files = 0
-            for (root, dirs, files) in walkdir(path)
-                for file_name in files
-                    if endswith(file_name, ".jl")
-                        node = parse_green_one(joinpath(root, file_name))
-                        _count_lines!(counts, node)
-                        n_files += 1
-                    end
-                end
-            end
-        end
-        push!(table, (; directory=dir, language=:Julia,
-                        sublanguage=nothing, files=n_files,
-                        code=counts["Code"],
-                        comments=counts["Comment"] + counts["Docstring"],
-                        blanks=counts["Blank"]))
-    end
-    return table
-end
